Move ActiveRecord related setup to the AR test file
