+++ spec/model.rb
@@ -34,48 +34,3 @@ class Comment < ActiveRecord::Base
   belongs_to :post
   belongs_to :user
 end
-
-class Student
-  attr_writer :name
-
-  def initialize(name, age)
-    @name = name
-    @age = age
-  end
-
-  def self.foo; end
-
-  def name
-    @name
-  end
-
-  def age
-    @age
-  end
-
-  def age=(age)
-    @age = age
-  end
-
-  def id=(id)
-    @id = id
-  end
-
-  def id
-    @id
-  end
-end
-
-class HighSchoolStudent < Student;end
-
-class School
-  def initialize(name)
-    @name = name
-  end
-
-  def name
-    @name
-  end
-end
-
-
